refactor: extract main.cpp into 4 modules (console, injector, cli, output-generator)
Break the 2,670-line main.cpp into focused modules:

- src/console.{hpp,cpp}: Console I/O, color helpers, banner, elevation check
- src/injector.{hpp,cpp}: PE parser, manual-map injection, process finder
- src/cli.{hpp,cpp}: Argument parsing, game selection menu, output selection
- src/output-generator.{hpp,cpp}: JSON parsing, text output, hierarchy,
  globals, entity paths, protobuf output

Deduplicates Collector/FlatCollector (3 copies) and TreeWriter (2 copies)
into single unified implementations. main.cpp is now ~500 lines of
high-level orchestration + live server code.

Pure refactoring -- no logic changes. Both targets build cleanly.
